// parser/types/supabase.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  public: {
    Tables: {
      cources: {
        Row: {
          course_name: string
          created_at: string
          description: string
          id: number
          origin_url: string
          price_starts_from: string | null
          tariff_price: Json | null
          type: Json
          updated_at: string
        }
        Insert: {
          course_name: string
          created_at?: string
          description?: string
          id?: number
          origin_url: string
          price_starts_from?: string | null
          tariff_price?: Json | null
          type: Json
          updated_at?: string
        }
        Update: {
          course_name?: string
          created_at?: string
          description?: string
          id?: number
          origin_url?: string
          price_starts_from?: string | null
          tariff_price?: Json | null
          type?: Json
          updated_at?: string
        }
        Relationships: []
      }
      leads: {
        Row: {
          created_at: string
          email: string | null
          id: string
          interest: string | null
          ip: string | null
          name: string
          phone: string | null
          utm: Json
        }
        Insert: {
          created_at?: string
          email?: string | null
          id?: string
          interest?: string | null
          ip?: string | null
          name: string
          phone?: string | null
          utm: Json
        }
        Update: {
          created_at?: string
          email?: string | null
          id?: string
          interest?: string | null
          ip?: string | null
          name?: string
          phone?: string | null
          utm?: Json
        }
        Relationships: []
      }
      parse_error: {
        Row: {
          client_error___msg: string | null
          course_title: string | null
          created_at: string
          error_msg: string
          id: number
          report_id: number
          type: string
          url: string | null
        }
        Insert: {
          client_error___msg?: string | null
          course_title?: string | null
          created_at?: string
          error_msg: string
          id?: number
          report_id: number
          type: string
          url?: string | null
        }
        Update: {
          client_error___msg?: string | null
          course_title?: string | null
          created_at?: string
          error_msg?: string
          id?: number
          report_id?: number
          type?: string
          url?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "parse_error_report_id_fkey"
            columns: ["report_id"]
            isOneToOne: true
            referencedRelation: "parse_report"
            referencedColumns: ["id"]
          },
        ]
      }
      parse_report: {
        Row: {
          avg_parse_time: number | null
          created_at: string
          duration: string | null
          id: number
          total_courses_parsed: number | null
          total_parse_time: number | null
          total_updated: number | null
        }
        Insert: {
          avg_parse_time?: number | null
          created_at?: string
          duration?: string | null
          id?: number
          total_courses_parsed?: number | null
          total_parse_time?: number | null
          total_updated?: number | null
        }
        Update: {
          avg_parse_time?: number | null
          created_at?: string
          duration?: string | null
          id?: number
          total_courses_parsed?: number | null
          total_parse_time?: number | null
          total_updated?: number | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DefaultSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof Database },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
  ? Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never